Extract the schema compile walk into free functions
The walk that turns a declaration into a validator (compile_node and the
_compile_dict/object/sequence/type/callable/literal/self helpers, _make_candidate,
the Self recursion machinery, and the leaf-validator factories) moves out of the
Schema class into src/probatio/_compile.py as free functions taking a small
CompileCtx (the extra and required policies, and a uses_self flag written back).

Schema keeps only its own concerns and calls compile_node; the walk is referenced
through the ctx, not self. The one Schema reference the walk still needs (an
isinstance check for a nested Schema reused as a value) is registered at import
through _compile._SCHEMA_CLS, so _compile never imports schema and the two stay
free of an import cycle. recursion_guard moves to _compile alongside the Self
machinery; its two consumers import it from there.

This decouples the compiler from the Schema class (which must stay a plain Python
class, since user code and the dataclass/TypedDict builders subclass it) and lets
the walk be accelerated as its own unit. Pure Python is perf-neutral: validation
is untouched, and construction stays within noise on realistic schemas.
